pyproject.toml
[project]
name = "spralingua"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
